requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
asgiref==3.2.10; python_version >= '3.5'
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
autobahn==20.7.1; python_version >= '3.5'
automat==20.2.0
cffi==1.14.2
channels==2.4.0
constantly==15.1.0
coverage==5.3
cryptography==3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
daphne==2.5.0
distlib==0.3.1
django-appconf==1.0.4
django==3.1.1; python_version >= '3.6'
filelock==3.0.12
hyperlink==20.0.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
incremental==17.5.0
iniconfig==1.0.1
more-itertools==8.5.0; python_version >= '3.5'
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyhamcrest==2.0.2; python_version >= '3.5'
pyopenssl==19.1.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytest-asyncio==0.14.0
pytest-cov==2.10.1
pytest-django==3.10.0
pytest==6.0.2
pytz==2020.1
service-identity==18.1.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
toml==0.10.1
tox==3.20.0
twisted[tls]==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
txaio==20.4.1; python_version >= '3.5'
virtualenv==20.0.31; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
zope.interface==5.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
